Extract non-DEM bake candidate factory
